LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/manual/filter/offscreencanvas.filter-expected.txt


FAIL none assert_array_equals: The image data generated by filter blur(10px) should be the same for both OffscreenCanvas and regular canvas expected property 0 to be 113 but got 0 (expected array object "113,38,28,27,111,34,26,30,108,39,23,33,106,43,28,36,105,46,26,39" got object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0")
